refactor: move column importers into their own module
