+++ pyprobe/cyclers/neware.py
@@ -5,24 +5,25 @@ import os
 
 import polars as pl
 
-from pyprobe.cyclers import basecycler as bc
+from pyprobe.cyclers import column_importers as ci
+from pyprobe.cyclers.basecycler import BaseCycler
 
 logger = logging.getLogger(__name__)
 
 
-class Neware(bc.BaseCycler):
+class Neware(BaseCycler):
     """A class to load and process Neware battery cycler data."""
 
-    column_importers: list[bc.ColumnMap] = [
-        bc.DateTime("Date", "%Y-%m-%d  %H:%M:%S%.f"),
-        bc.TimeFromDate("Date", "%Y-%m-%d  %H:%M:%S%.f"),
-        bc.CastAndRename("Step", "Step Index", pl.Int64),
-        bc.ConvertUnits("Current [A]", "Current(*)"),
-        bc.ConvertUnits("Voltage [V]", "Voltage(*)"),
-        bc.CapacityFromChDch("Chg. Cap.(*)", "DChg. Cap.(*)"),
-        bc.ConvertTemperature("T1(*)"),
-        bc.TimeFromDate("Total Time", "%Y-%m-%d  %H:%M:%S%.f"),
-        bc.CapacityFromCurrentSign("Capacity(*)", "Current(*)"),
+    column_importers: list[ci.ColumnMap] = [
+        ci.DateTime("Date", "%Y-%m-%d  %H:%M:%S%.f"),
+        ci.TimeFromDate("Date", "%Y-%m-%d  %H:%M:%S%.f"),
+        ci.CastAndRename("Step", "Step Index", pl.Int64),
+        ci.ConvertUnits("Current [A]", "Current(*)"),
+        ci.ConvertUnits("Voltage [V]", "Voltage(*)"),
+        ci.CapacityFromChDch("Chg. Cap.(*)", "DChg. Cap.(*)"),
+        ci.ConvertTemperature("T1(*)"),
+        ci.TimeFromDate("Total Time", "%Y-%m-%d  %H:%M:%S%.f"),
+        ci.CapacityFromCurrentSign("Capacity(*)", "Current(*)"),
     ]
 
     @staticmethod
