+++ tests/test_cell.py
@@ -15,7 +15,7 @@ from polars.testing import assert_frame_equal
 import pyprobe
 from pyprobe import cell
 from pyprobe._version import __version__
-from pyprobe.cyclers import basecycler
+from pyprobe.cyclers import column_importers as ci
 from pyprobe.readme_processor import process_readme
 
 
@@ -119,11 +119,11 @@ def test_process_generic_file(cell_instance):
     )
 
     column_importers = [
-        basecycler.ConvertUnits("Time [s]", "T [*]"),
-        basecycler.ConvertUnits("Voltage [V]", "V [*]"),
-        basecycler.ConvertUnits("Current [A]", "I [*]"),
-        basecycler.ConvertUnits("Capacity [Ah]", "Q [*]"),
-        basecycler.CastAndRename("Step", "Count", pl.Int64),
+        ci.ConvertUnits("Time [s]", "T [*]"),
+        ci.ConvertUnits("Voltage [V]", "V [*]"),
+        ci.ConvertUnits("Current [A]", "I [*]"),
+        ci.ConvertUnits("Capacity [Ah]", "Q [*]"),
+        ci.CastAndRename("Step", "Count", pl.Int64),
     ]
 
     df.write_csv(f"{folder_path}/test_generic_file.csv")
@@ -674,7 +674,7 @@ def test_import_from_cycler(cell_instance, mocker):
     )
 
     # Test with column_importers provided
-    column_importers = [basecycler.ConvertUnits("Time [s]", "T [*]")]
+    column_importers = [ci.ConvertUnits("Time [s]", "T [*]")]
     cell_instance.import_from_cycler(
         procedure_name,
         cycler,
@@ -694,7 +694,7 @@ def test_import_from_cycler(cell_instance, mocker):
     )
 
     # Test with extra cycler columns provided
-    extra_cycler_columns = [basecycler.ConvertUnits("Time [s]", "T [*]")]
+    extra_cycler_columns = [ci.ConvertUnits("Time [s]", "T [*]")]
     cell_instance.import_from_cycler(
         procedure_name,
         cycler,
@@ -774,11 +774,11 @@ def test_process_cycler_data_generic():
     )
 
     column_importers = [
-        basecycler.ConvertUnits("Time [s]", "T [*]"),
-        basecycler.ConvertUnits("Voltage [V]", "V [*]"),
-        basecycler.ConvertUnits("Current [A]", "I [*]"),
-        basecycler.ConvertUnits("Capacity [Ah]", "Q [*]"),
-        basecycler.CastAndRename("Step", "Count", pl.Int64),
+        ci.ConvertUnits("Time [s]", "T [*]"),
+        ci.ConvertUnits("Voltage [V]", "V [*]"),
+        ci.ConvertUnits("Current [A]", "I [*]"),
+        ci.ConvertUnits("Capacity [Ah]", "Q [*]"),
+        ci.CastAndRename("Step", "Count", pl.Int64),
     ]
 
     df.write_csv(data_path)
